Refactor expired_project_cleanup
